requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
boto3==1.16.21
botocore==1.19.21
celery==5.0.2
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.0.5
django-bmemcached==0.3.0
django-cors-headers==3.5.0
django-environ==0.4.5
django-storages==1.10.1
djangorestframework==3.12.2
djongo==1.3.3
gunicorn==20.0.4
importlib-metadata==2.0.0
jmespath==0.10.0
kombu==5.0.2
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
PyJWT==1.7.1
pymongo==3.11.1
python-binary-memcached==0.30.1
python-dateutil==2.8.1
pytz==2020.4
redis==3.5.3
rq==1.6.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.2.4
uhashring==1.2
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
